requirements.txt
jittor
matplotlib>=3.5.0,<3.8.0
numpy>=1.21.0,<1.25.0
Pillow>=9.0.0,<11.0.0
PyYAML>=6.0.0
scikit-learn>=1.0.0,<1.4.0
scikit-image>=0.18.0,<0.26.0
scipy>=1.7.0,<1.15.0
tqdm>=4.60.0
wandb>=0.15.0
opencv-python>=4.5.0